refactor(api): :recycle: Increase robustness in api
Fixes #2057
