+++ api/package.json
@@ -3,7 +3,7 @@
   "license": "MIT",
   "homepage": "https://openarabic.io",
   "repository": "https://github.com/edenmind/OpenArabic",
-  "version": "1444.4.179",
+  "version": "1444.4.180",
   "authors": [
     "Yunus Andreasson <yunus@edenmind.com> (https://github.com/YunusAndreasson)"
   ],
@@ -20,7 +20,7 @@
     "lint": "eslint '**/*.js' --ignore-pattern node_modules/"
   },
   "dependencies": {
-    "@aws-sdk/client-s3": "^3.306.0",
+    "@aws-sdk/client-s3": "^3.309.0",
     "@fastify/autoload": "^5.7.1",
     "@fastify/cors": "^8.2.1",
     "@fastify/env": "^4.2.0",
@@ -44,14 +44,14 @@
     "uuid": "^9.0.0"
   },
   "devDependencies": {
-    "eslint": "^8.37.0",
+    "eslint": "^8.38.0",
     "eslint-plugin-import": "^2.27.5",
     "eslint-plugin-prettier": "^4.2.1",
-    "eslint-plugin-putout": "^17.2.1",
+    "eslint-plugin-putout": "^17.3.0",
     "eslint-plugin-security": "^1.7.1",
     "eslint-plugin-unicorn": "^46.0.0",
     "prettier": "^2.8.7",
-    "putout": "^29.2.1",
+    "putout": "^29.2.3",
     "tap": "^16.3.4"
   }
 }
