+++ package/templates/api-hpa.yaml
@@ -8,7 +8,7 @@ spec:
     kind: Deployment
     name: api
   minReplicas: 2
-  maxReplicas: 5
+  maxReplicas: 3
   metrics:
     - type: Resource
       resource:
